Fixed pixel size bug, changing HSize and VSize to integers made the division of them an int division instead of double division.

+++ TheRayTracerChallenge/Camera.cs
@@ -27,7 +27,7 @@ namespace TheRayTracerChallenge
         private void PrecalculateAndSetCanvasData()
         {
             var halfView = Math.Tan(FieldOfView / 2);
-            var aspect = HSize / VSize;
+            var aspect = (double)HSize / VSize;
 
             if (aspect >= 1)
             {
